refactor: split docs graph tests
