+++ crates/ox_content_docs/src/graph.rs
@@ -1122,575 +1122,4 @@ fn normalize_existing_path(path: &Path) -> PathBuf {
 }
 
 #[cfg(test)]
-mod tests {
-    use std::fs;
-    use std::sync::atomic::{AtomicU64, Ordering};
-    use std::time::{SystemTime, UNIX_EPOCH};
-
-    use super::*;
-
-    #[test]
-    fn builds_export_graph_and_extracts_entrypoint_docs() {
-        let root = temp_root();
-        fs::create_dir_all(root.join("src")).unwrap();
-        fs::write(
-            root.join("src/index.ts"),
-            r"
-export { add as sum } from './math';
-export type { Options } from './types';
-export * from './extra';
-export { ExternalThing } from 'external-pkg/subpath';
-",
-        )
-        .unwrap();
-        fs::write(
-            root.join("src/math.ts"),
-            r"
-/** Adds two numbers. */
-export function add(a: number, b: number): number {
-  return a + b;
-}
-",
-        )
-        .unwrap();
-        fs::write(
-            root.join("src/types.ts"),
-            r"
-/** Runtime options. */
-export interface Options {
-  value: string;
-}
-",
-        )
-        .unwrap();
-        fs::write(
-            root.join("src/extra.ts"),
-            r"
-/** Creates a label. */
-export function label(value: string): string {
-  return value;
-}
-",
-        )
-        .unwrap();
-
-        let entrypoints = [EntryPointSpec {
-            path: PathBuf::from("src/index.ts"),
-            name: Some("default".to_string()),
-        }];
-        let graph_options = GraphOptions { root: Some(root.clone()), ..GraphOptions::default() };
-
-        let graph = build_export_graph(&entrypoints, &graph_options).unwrap();
-        assert_eq!(graph.entrypoints[0].name, "default");
-        assert!(graph.entrypoints[0].exports.iter().any(|export| export.name == "sum"));
-        assert!(graph.entrypoints[0].exports.iter().any(|export| export.name == "Options"));
-        assert!(graph.entrypoints[0].exports.iter().any(|export| export.name == "label"));
-        assert!(graph.entrypoints[0].exports.iter().any(|export| matches!(
-            &export.source,
-            ExportSource::External { package, original_name, .. }
-                if package == "external-pkg" && original_name == "ExternalThing"
-        )));
-
-        let docs = extract_docs_from_entry_points(
-            &entrypoints,
-            &EntryPointDocsOptions {
-                graph: graph_options,
-                include_private: false,
-                include_internal: false,
-                type_parameters: false,
-            },
-        )
-        .unwrap();
-        let names = docs[0].entries.iter().map(|entry| entry.name.as_str()).collect::<Vec<_>>();
-        assert_eq!(names, ["sum", "Options", "label"]);
-
-        fs::remove_dir_all(root).unwrap();
-    }
-
-    #[test]
-    fn entrypoint_docs_capture_module_level_description() {
-        let root = temp_root();
-        fs::create_dir_all(root.join("src")).unwrap();
-        // Entry file with a module-level `@module` summary that only re-exports.
-        fs::write(
-            root.join("src/context.ts"),
-            r"
-/**
- * The entry for gunshi context.
- *
- * @example
- * ```ts
- * createCommandContext()
- * ```
- *
- * @experimental This entry point is experimental.
- *
- * @module
- */
-export { createCommandContext } from './context-impl';
-",
-        )
-        .unwrap();
-        fs::write(
-            root.join("src/context-impl.ts"),
-            r"
-/** Creates a command context. */
-export function createCommandContext(): void {}
-",
-        )
-        .unwrap();
-        // Entry file without any module-level comment.
-        fs::write(
-            root.join("src/plugin.ts"),
-            r"
-export { plugin } from './plugin-impl';
-",
-        )
-        .unwrap();
-        fs::write(
-            root.join("src/plugin-impl.ts"),
-            r"
-/** Defines a plugin. */
-export function plugin(): void {}
-",
-        )
-        .unwrap();
-
-        let entrypoints = [
-            EntryPointSpec {
-                path: PathBuf::from("src/context.ts"),
-                name: Some("context".to_string()),
-            },
-            EntryPointSpec {
-                path: PathBuf::from("src/plugin.ts"),
-                name: Some("plugin".to_string()),
-            },
-        ];
-        let graph_options = GraphOptions { root: Some(root.clone()), ..GraphOptions::default() };
-
-        let docs = extract_docs_from_entry_points(
-            &entrypoints,
-            &EntryPointDocsOptions {
-                graph: graph_options,
-                include_private: false,
-                include_internal: false,
-                type_parameters: false,
-            },
-        )
-        .unwrap();
-
-        let context = docs.iter().find(|module| module.name == "context").unwrap();
-        assert_eq!(context.file, "context");
-        assert_eq!(context.description, "The entry for gunshi context.");
-        assert_eq!(context.examples, vec!["```ts\ncreateCommandContext()\n```".to_string()]);
-        assert_eq!(context.tags.len(), 1);
-        assert_eq!(context.tags[0].tag, "experimental");
-        assert_eq!(context.tags[0].value, "This entry point is experimental.");
-        // The module entry itself is not surfaced as a regular export entry.
-        assert!(context.entries.iter().all(|entry| entry.kind != NormalizedDocKind::Module));
-        assert!(context.entries.iter().any(|entry| entry.name == "createCommandContext"));
-
-        let plugin = docs.iter().find(|module| module.name == "plugin").unwrap();
-        assert_eq!(plugin.file, "plugin");
-        assert!(plugin.description.is_empty());
-
-        fs::remove_dir_all(root).unwrap();
-    }
-
-    #[test]
-    fn entrypoint_docs_prefers_explicit_module_name() {
-        let root = temp_root();
-        fs::create_dir_all(root.join("src")).unwrap();
-        fs::write(
-            root.join("src/index.ts"),
-            r"
-/**
- * Default entry point.
- *
- * @module default
- */
-/** Runs the CLI. */
-export function cli(): void {}
-",
-        )
-        .unwrap();
-
-        let docs = extract_docs_from_entry_points(
-            &[EntryPointSpec {
-                path: PathBuf::from("src/index.ts"),
-                name: Some("entry".to_string()),
-            }],
-            &EntryPointDocsOptions {
-                graph: GraphOptions { root: Some(root.clone()), ..GraphOptions::default() },
-                include_private: false,
-                include_internal: false,
-                type_parameters: false,
-            },
-        )
-        .unwrap();
-
-        assert_eq!(docs[0].name, "default");
-        assert_eq!(docs[0].file, "default");
-        assert_eq!(docs[0].description, "Default entry point.");
-
-        fs::remove_dir_all(root).unwrap();
-    }
-
-    #[test]
-    fn entrypoint_docs_uses_entrypoint_name_without_module_tag() {
-        let root = temp_root();
-        fs::create_dir_all(root.join("src")).unwrap();
-        fs::write(
-            root.join("src/agent.ts"),
-            r"
-export function isAiAgent(): boolean {
-  return false;
-}
-",
-        )
-        .unwrap();
-
-        let docs = extract_docs_from_entry_points(
-            &[EntryPointSpec {
-                path: PathBuf::from("src/agent.ts"),
-                name: Some("agent".to_string()),
-            }],
-            &EntryPointDocsOptions {
-                graph: GraphOptions { root: Some(root.clone()), ..GraphOptions::default() },
-                include_private: false,
-                include_internal: false,
-                type_parameters: false,
-            },
-        )
-        .unwrap();
-
-        assert_eq!(docs[0].name, "agent");
-        assert_eq!(docs[0].file, "agent");
-        assert!(docs[0].description.is_empty());
-
-        fs::remove_dir_all(root).unwrap();
-    }
-
-    #[test]
-    fn entrypoint_docs_emit_public_consts_without_jsdoc() {
-        let root = temp_root();
-        fs::create_dir_all(root.join("src")).unwrap();
-        fs::write(
-            root.join("src/index.ts"),
-            r"
-export { ANONYMOUS_COMMAND_NAME, CLI_OPTIONS_DEFAULT } from './constants';
-",
-        )
-        .unwrap();
-        fs::write(
-            root.join("src/constants.ts"),
-            r#"
-export const ANONYMOUS_COMMAND_NAME = "(anonymous)";
-
-export const CLI_OPTIONS_DEFAULT: CliOptions<DefaultGunshiParams> = {
-  usageSilent: false,
-};
-"#,
-        )
-        .unwrap();
-
-        let entrypoints = [EntryPointSpec {
-            path: PathBuf::from("src/index.ts"),
-            name: Some("default".to_string()),
-        }];
-        let docs = extract_docs_from_entry_points(
-            &entrypoints,
-            &EntryPointDocsOptions {
-                graph: GraphOptions { root: Some(root.clone()), ..GraphOptions::default() },
-                include_private: false,
-                include_internal: false,
-                type_parameters: false,
-            },
-        )
-        .unwrap();
-
-        assert!(docs[0].diagnostics.is_empty());
-        let anonymous =
-            docs[0].entries.iter().find(|entry| entry.name == "ANONYMOUS_COMMAND_NAME").unwrap();
-        assert_eq!(anonymous.kind.as_str(), "variable");
-        assert!(anonymous.description.is_empty());
-        assert_eq!(
-            anonymous.signature.as_deref(),
-            Some(r#"export const ANONYMOUS_COMMAND_NAME = "(anonymous)""#)
-        );
-
-        let cli_options =
-            docs[0].entries.iter().find(|entry| entry.name == "CLI_OPTIONS_DEFAULT").unwrap();
-        assert_eq!(cli_options.kind.as_str(), "variable");
-        assert_eq!(
-            cli_options.signature.as_deref(),
-            Some("export const CLI_OPTIONS_DEFAULT: CliOptions<DefaultGunshiParams>")
-        );
-
-        fs::remove_dir_all(root).unwrap();
-    }
-
-    #[test]
-    fn entrypoint_docs_diagnose_internal_type_filtered_by_visibility() {
-        let root = temp_root();
-        fs::create_dir_all(root.join("src")).unwrap();
-        fs::write(root.join("src/index.ts"), "export type { ExtractArgs } from './types';\n")
-            .unwrap();
-        fs::write(
-            root.join("src/types.ts"),
-            r"
-/**
- * Type helper to extract args.
- *
- * @internal
- */
-export type ExtractArgs<G> = G extends { args: infer A } ? A : never;
-",
-        )
-        .unwrap();
-
-        let entrypoints = [EntryPointSpec {
-            path: PathBuf::from("src/index.ts"),
-            name: Some("default".to_string()),
-        }];
-        let docs = extract_docs_from_entry_points(
-            &entrypoints,
-            &EntryPointDocsOptions {
-                graph: GraphOptions { root: Some(root.clone()), ..GraphOptions::default() },
-                include_private: false,
-                include_internal: false,
-                type_parameters: false,
-            },
-        )
-        .unwrap();
-
-        assert!(docs[0].entries.is_empty());
-        assert_eq!(docs[0].diagnostics.len(), 1);
-        assert_eq!(docs[0].diagnostics[0].code, DocsDiagnosticCode::FilteredByVisibility);
-        assert_eq!(docs[0].diagnostics[0].export_name, "ExtractArgs");
-        assert!(docs[0].diagnostics[0].message.contains("@internal"));
-
-        let with_internal = extract_docs_from_entry_points(
-            &entrypoints,
-            &EntryPointDocsOptions {
-                graph: GraphOptions { root: Some(root.clone()), ..GraphOptions::default() },
-                include_private: false,
-                include_internal: true,
-                type_parameters: false,
-            },
-        )
-        .unwrap();
-        assert_eq!(with_internal[0].entries[0].name, "ExtractArgs");
-        assert!(with_internal[0].diagnostics.is_empty());
-
-        fs::remove_dir_all(root).unwrap();
-    }
-
-    #[test]
-    fn extracts_docs_from_resolved_external_package_types() {
-        let root = temp_root();
-        fs::create_dir_all(root.join("src")).unwrap();
-        fs::write(root.join("src/index.ts"), "export { ExternalThing } from 'external-pkg';\n")
-            .unwrap();
-        write_external_package(
-            &root,
-            "external-pkg",
-            r"
-/** External thing. */
-export interface ExternalThing {
-  value: string;
-}
-",
-        );
-
-        let entrypoints = [EntryPointSpec {
-            path: PathBuf::from("src/index.ts"),
-            name: Some("default".to_string()),
-        }];
-        let graph_options = GraphOptions {
-            root: Some(root.clone()),
-            external_docs: ExternalDocsOptions { enabled: true, package_sources: vec![] },
-            ..GraphOptions::default()
-        };
-
-        let graph = build_export_graph(&entrypoints, &graph_options).unwrap();
-        assert!(graph.entrypoints[0].exports.iter().any(|export| matches!(
-            &export.source,
-            ExportSource::External { package, module: Some(module), original_name, .. }
-                if package == "external-pkg"
-                    && original_name == "ExternalThing"
-                    && module.ends_with("index.d.ts")
-        )));
-
-        let docs = extract_docs_from_entry_points(
-            &entrypoints,
-            &EntryPointDocsOptions {
-                graph: graph_options,
-                include_private: false,
-                include_internal: false,
-                type_parameters: false,
-            },
-        )
-        .unwrap();
-
-        assert_eq!(docs[0].entries[0].name, "ExternalThing");
-        assert_eq!(docs[0].entries[0].description, "External thing.");
-
-        fs::remove_dir_all(root).unwrap();
-    }
-
-    #[test]
-    fn follows_import_aliases_in_declaration_barrels() {
-        let root = temp_root();
-        fs::create_dir_all(root.join("src")).unwrap();
-        fs::write(root.join("src/index.ts"), "export { parseArgs } from 'external-pkg';\n")
-            .unwrap();
-        let package_root = root.join("node_modules/external-pkg");
-        fs::create_dir_all(package_root.join("lib")).unwrap();
-        fs::write(
-            package_root.join("package.json"),
-            r#"{
-  "name": "external-pkg",
-  "type": "module",
-  "exports": {
-    ".": {
-      "types": "./lib/index.d.ts",
-      "default": "./lib/index.js"
-    }
-  }
-}"#,
-        )
-        .unwrap();
-        fs::write(
-            package_root.join("lib/index.d.ts"),
-            r#"
-import { a as parseArgs } from "./parser-hash.js";
-export { parseArgs };
-"#,
-        )
-        .unwrap();
-        fs::write(
-            package_root.join("lib/parser-hash.d.ts"),
-            r"
-/** Parse args. */
-declare function a(): void;
-export { a };
-",
-        )
-        .unwrap();
-
-        let entrypoints = [EntryPointSpec {
-            path: PathBuf::from("src/index.ts"),
-            name: Some("default".to_string()),
-        }];
-        let graph_options = GraphOptions {
-            root: Some(root.clone()),
-            external_docs: ExternalDocsOptions { enabled: true, package_sources: vec![] },
-            ..GraphOptions::default()
-        };
-        let docs = extract_docs_from_entry_points(
-            &entrypoints,
-            &EntryPointDocsOptions {
-                graph: graph_options,
-                include_private: false,
-                include_internal: false,
-                type_parameters: false,
-            },
-        )
-        .unwrap();
-
-        assert_eq!(docs[0].entries[0].name, "parseArgs");
-        assert_eq!(docs[0].entries[0].description, "Parse args.");
-        // The alias resolved to a declaration under node_modules, so its absolute
-        // source path is dropped (no "View source" link, no local-path leak).
-        assert!(docs[0].entries[0].file.is_empty());
-
-        fs::remove_dir_all(root).unwrap();
-    }
-
-    #[test]
-    fn external_package_source_mapping_prefers_workspace_source() {
-        let root = temp_root();
-        fs::create_dir_all(root.join("src")).unwrap();
-        fs::create_dir_all(root.join("packages/plugin/src")).unwrap();
-        fs::write(root.join("src/index.ts"), "export { helper } from '@scope/plugin';\n").unwrap();
-        fs::write(
-            root.join("packages/plugin/src/index.ts"),
-            r"
-/** Workspace helper. */
-export function helper(): void {}
-",
-        )
-        .unwrap();
-
-        let entrypoints = [EntryPointSpec {
-            path: PathBuf::from("src/index.ts"),
-            name: Some("default".to_string()),
-        }];
-        let graph_options = GraphOptions {
-            root: Some(root.clone()),
-            external_docs: ExternalDocsOptions {
-                enabled: true,
-                package_sources: vec![ExternalPackageSource {
-                    package: "@scope/plugin".to_string(),
-                    entry: PathBuf::from("packages/plugin/src/index.ts"),
-                }],
-            },
-            ..GraphOptions::default()
-        };
-
-        let docs = extract_docs_from_entry_points(
-            &entrypoints,
-            &EntryPointDocsOptions {
-                graph: graph_options,
-                include_private: false,
-                include_internal: false,
-                type_parameters: false,
-            },
-        )
-        .unwrap();
-
-        assert_eq!(docs[0].entries[0].name, "helper");
-        assert_eq!(docs[0].entries[0].description, "Workspace helper.");
-        assert!(docs[0].entries[0].file.ends_with("packages/plugin/src/index.ts"));
-
-        fs::remove_dir_all(root).unwrap();
-    }
-
-    fn temp_root() -> PathBuf {
-        static COUNTER: AtomicU64 = AtomicU64::new(0);
-        let nanos = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_nanos();
-        let seq = COUNTER.fetch_add(1, Ordering::Relaxed);
-        let mut dirname =
-            StringBuilder::with_capacity("ox-content-docs-graph-".len() + 39 + 1 + 20);
-        dirname.push_str("ox-content-docs-graph-");
-        dirname.push_u128(nanos);
-        dirname.push_char('-');
-        dirname.push_u128(u128::from(seq));
-        std::env::temp_dir().join(dirname.into_string())
-    }
-
-    fn write_external_package(root: &Path, name: &str, declaration: &str) {
-        let package_root = root.join("node_modules").join(name);
-        fs::create_dir_all(package_root.join("lib")).unwrap();
-        let mut package_json = StringBuilder::with_capacity(150 + name.len());
-        package_json.push_str(
-            r#"{
-  "name": ""#,
-        );
-        package_json.push_str(name);
-        package_json.push_str(
-            r#"",
-  "type": "module",
-  "exports": {
-    ".": {
-      "types": "./lib/index.d.ts",
-      "default": "./lib/index.js"
-    }
-  }
-}"#,
-        );
-        fs::write(package_root.join("package.json"), package_json.into_string()).unwrap();
-        fs::write(package_root.join("lib/index.d.ts"), declaration).unwrap();
-    }
-}
+mod tests;
