+++ crates/ox_content_docs/src/graph/tests.rs
@@ -1,0 +1,4 @@
+mod entrypoints;
+mod external;
+mod modules;
+mod support;
